+++ frontend/src/components/FrameComponent.tsx
@@ -1,9 +1,9 @@
 import { FunctionComponent, useCallback } from "react";
 import { useNavigate } from "react-router-dom";
 import NewRx from "./NewRx";
-import NewPtSearch from "./NewPtSearch";
-import NewDrSearch from "./NewDrSearch";
-import RxItemSearch from "./RxItemSearch";
+import NewPtSearch from "../pages/NewPtSearch";
+import NewDrSearch from "../pages/NewDrSearch";
+import RxItemSearch from "../pages/RxItemSearch";
 import Search from "./Search";
 
 export type FrameComponentType = {
